[C++] FFModel::conv2d: groups argument should be after activation (#118)
This makes the C++ API consistent with the Python API. It also fixes
bugs in many of the C++ tests. Many of the C++ tests were passing
an activation for the groups argument. Not sure how they worked. The code
would still compile because activation is an enum and groups is an int.